==> MarkDown/17b.PvsNP.cc.Rmd <==
---
title: "P vs NP, differential measurement"
author: "HBG"
date: "11/5/2018"
output: pdf_document
---
#This script use the differential measure (z-score) between tumor/non-tumor cells
#
```{r}
### this block perform z-score calculations
library("microbenchmark")
library("matrixStats")

## the GO terms for cc
GOcategory.file <- read.csv("../Data/human.cc.term.csv",header=TRUE, stringsAsFactors=F)
cc.go.cat <- GOcategory.file$GO.term
cc.dim <- length(cc.go.cat)
conn.dim <- 20
hspin <- matrix(as.numeric(unlist(read.table("human.rnf43.PvNP.cc.txt", header=F, sep=","))), nrow=cc.dim, ncol=conn.dim)
obs <- c(hspin)

perm <- c()
for (i in 1:100) {
    name <- paste("ms02.human", "/", "rnf43.PvNP", "/", "ms02.", i, ".cc.matrix.csv", sep="")
    mat <- matrix(as.numeric(unlist(read.table(name, header=F, sep=","))), nrow=cc.dim, ncol=conn.dim)
    perm <- rbind(perm, c(mat))
}

mean <- colMeans(perm)
std <- colSds(perm)

zscore <- round((obs - mean)/std, 3)

z.pvnp <- matrix(zscore, nrow=cc.dim, ncol=conn.dim)

#write.table(z, file="human.rnf43.PvNP.cc.z.csv", sep=",", row.names=F, col.names=F, quote=F)
```

```{r}
library('gplots')
z.pvnp <- t(z.pvnp)

rownames(z.pvnp) <- c("5%","10%","15%","20%","25%","30%","35%","40%","45%","50%",
                 "55%","60%","65%","70%","75%","80%","85%","90%","95%","100%")

colnames(z.pvnp) <- cc.go.cat

colors = c(seq(min(z.pvnp),-10.1,length=100),seq(-9.9,9.9,length=100),seq(10.1,max(z.pvnp),length=100))
my_palette <- colorRampPalette(c("blue2", "white", "red2"))(n = 299)

png(filename = "17.PvsNP.z.hm.png",width=6, height=5.5, res=1200, unit="in")
heatmap.2(z.pvnp, col=my_palette, breaks=colors, dendrogram='none', Rowv=F, Colv=F,
          trace='none', offsetRow = 0, offsetCol = 0,
          xlab="Biological Process Terms", ylab="PDAC vs Normal Cells, Quantiles",
          margins = c(2,3.5), key.title = "Color Bar", key.xlab="Z-score", key.ylab=NA,
          revC = T,
          labCol = NA, #labRow =,
          #srtCol=45, adjCol=c(1,0),
          #lmat=rbind(c(0,3,4), c(2,1,0)), lwid=c(1.5,4,2),
          scale="none", symbreaks=T, symm=F, symkey = F)
dev.off()

```

```{r}
#correlations to quantile 20 (95%-100%) genes
y.val <- c()
for (i in 1:20) {
  y.val2 <- c()
  for (j in 1:20) {
     cor.val <- cor.test(z.pvnp[j,], z.pvnp[i,])
     y.val2 <- rbind(y.val2, cor.val$estimate)
  }
  y.val <- cbind(y.val, y.val2)
}

rownames(y.val) <- c("5%","10%","15%","20%","25%","30%","35%","40%","45%","50%",
                   "55%","60%","65%","70%","75%","80%","85%","90%","95%","100%")
colnames(y.val) <- c("5%","10%","15%","20%","25%","30%","35%","40%","45%","50%",
                   "55%","60%","65%","70%","75%","80%","85%","90%","95%","100%")
my_palette2 <- colorRampPalette(c("blue2", "white", "red2"))(9)
#png(filename = "17.PvsNP.z.hm.cor.matrix.png",width=6, height=5.5, res=1200, unit="in")
heatmap.2(y.val, col=my_palette2, dendrogram='none', Rowv=F, Colv=F,
          trace='none', offsetRow = 0, offsetCol = 0,
          xlab="Quantiles of Bayes Factors", ylab="Quantiles of Bayes Factors",
          margins = c(3.5,3.5), key.title = "Color Bar", key.xlab="Pearson Correlation", key.ylab=NA,
          revC = T,
          #labCol = NA, #labRow =,
          #srtCol=45, adjCol=c(1,0),
          #lmat=rbind(c(0,3,4), c(2,1,0)), lwid=c(1.5,4,2),
          scale="none", symbreaks=T, symm=F, symkey = F)
#dev.off()

```

```{r}
library(GO.db)

hspin <- matrix(as.numeric(unlist(read.table("human.rnf43.all.cc.txt", header=F, sep=","))), nrow=cc.dim, ncol=conn.dim)
obs <- c(hspin)

perm <- c()
for (i in 1:100) {
    name <- paste("ms02.human", "/", "rnf43.heatmap", "/", "ms02.", i, ".cc.matrix.csv", sep="")
    mat <- matrix(as.numeric(unlist(read.table(name, header=F, sep=","))), nrow=cc.dim, ncol=conn.dim)
    perm <- rbind(perm, c(mat))
}

mean <- colMeans(perm)
std <- colSds(perm)

zscore <- round((obs - mean)/std, 3)

z.p <- matrix(zscore, nrow=cc.dim, ncol=conn.dim)
z.p <- t(z.p)
rownames(z.p) <- c("5%","10%","15%","20%","25%","30%","35%","40%","45%","50%",
                 "55%","60%","65%","70%","75%","80%","85%","90%","95%","100%")
colnames(z.p) <- cc.go.cat

hspin <- matrix(as.numeric(unlist(read.table("human.rnf43.NP.cc.txt", header=F, sep=","))), nrow=cc.dim, ncol=conn.dim)
obs <- c(hspin)

perm <- c()
for (i in 1:100) {
    name <- paste("ms02.human", "/", "rnf43.NP", "/", "ms02.", i, ".cc.matrix.csv", sep="")
    mat <- matrix(as.numeric(unlist(read.table(name, header=F, sep=","))), nrow=cc.dim, ncol=conn.dim)
    perm <- rbind(perm, c(mat))
}

mean <- colMeans(perm)
std <- colSds(perm)

zscore <- round((obs - mean)/std, 3)

z.np <- matrix(zscore, nrow=cc.dim, ncol=conn.dim)
z.np <- t(z.np)
rownames(z.np) <- c("5%","10%","15%","20%","25%","30%","35%","40%","45%","50%",
                 "55%","60%","65%","70%","75%","80%","85%","90%","95%","100%")
colnames(z.np) <- cc.go.cat

enrich <- colnames(z.pvnp[, which(z.pvnp[20,] > 4 & z.pvnp[1,] < z.pvnp[20,]/2 & z.pvnp[2,] < z.pvnp[20,]/2 & 
                                  z.pvnp[3,] < z.pvnp[20,]/2 & z.pvnp[4,] < z.pvnp[20,]/2 & 
                                  z.pvnp[5,] < z.pvnp[20,]/2 & z.pvnp[6,] < z.pvnp[20,]/2 & 
                                  z.pvnp[7,] < z.pvnp[20,]/2 & z.pvnp[8,] < z.pvnp[20,]/2 & 
                                  z.pvnp[9,] < z.pvnp[20,]/2 & z.pvnp[10,] < z.pvnp[20,]/2 &
                                  z.pvnp[11,] < z.pvnp[20,]-2 & z.pvnp[12,] < z.pvnp[20,]-2 &
                                  z.pvnp[13,] < z.pvnp[20,]-2 & z.pvnp[14,] < z.pvnp[20,]-2 &
                                  z.pvnp[15,] < z.pvnp[20,]-2 & z.pvnp[16,] < z.pvnp[20,]-2 & 
                                  z.pvnp[17,] < z.pvnp[20,]-2 & z.pvnp[18,] < z.pvnp[20,]-2 )])
suppress <- colnames(z.pvnp[, which(z.pvnp[20,] < -4 & z.pvnp[1,] > z.pvnp[20,]/2 & z.pvnp[2,] > z.pvnp[20,]/2 & 
                                  z.pvnp[3,] > z.pvnp[20,]/2 & z.pvnp[4,] > z.pvnp[20,]/2 & 
                                  z.pvnp[5,] > z.pvnp[20,]/2 & z.pvnp[6,] > z.pvnp[20,]/2 & 
                                  z.pvnp[7,] > z.pvnp[20,]/2 & z.pvnp[8,] > z.pvnp[20,]/2 & 
                                  z.pvnp[9,] > z.pvnp[20,]/2 & z.pvnp[10,] > z.pvnp[20,]/2 &
                                  z.pvnp[11,] > z.pvnp[20,] & z.pvnp[12,] > z.pvnp[20,] &
                                  z.pvnp[13,] > z.pvnp[20,] & z.pvnp[14,] > z.pvnp[20,] &
                                  z.pvnp[15,] > z.pvnp[20,] & z.pvnp[16,] > z.pvnp[20,] & 
                                  z.pvnp[17,] > z.pvnp[20,] & z.pvnp[18,] > z.pvnp[20,] )])

enriched.terms <- c("GO.ID", "GO.Term", "Z-score")
for (i in 1:length(enrich)) {
  id <- as.character(enrich[i])
  term <- Term(GOID(id))
  seri <- which(cc.go.cat %in% id)
  #z.tumor <- z.p[20,seri]
  #z.normal <- z.np[20,seri]
  z.diff <- z.pvnp[20,seri]
  enriched.terms <- rbind(enriched.terms, c(id, term, z.diff))
}
enriched.terms
write.table(enriched.terms, file="17.PvNP.cc.enriched.csv", sep="\t", col.names = F, row.names = F, quote=F)

suppressed.terms <- c("GO.ID", "GO.Term", "Z.Differential")
for (i in 1:length(suppress)) {
  id <- as.character(suppress[i])
  term <- Term(GOID(id))
  seri <- which(cc.go.cat %in% id)
  #z.tumor <- z.p[20,seri]
  #z.normal <- z.np[20,seri]
  z.diff <- z.pvnp[20,seri]
  suppressed.terms <- rbind(suppressed.terms, c(id, term, z.diff))
}
suppressed.terms
write.table(suppressed.terms, file="17.PvNP.cc.suppressed.csv", sep="\t", col.names = F, row.names = F, quote=F)
  
```

```{r}
# PDAC cells

p.pos <- z.p[, which(z.p[20,] > 20)]
heatmap.2(p.pos, col=my_palette, dendrogram='none',
          trace='none', Rowv=F, Colv=F, revC=T, #labCol=NA,
          xlab="Cellular Composition Terms", ylab="Quantiles of Tumor Cells Bayes Factors",
          margins = c(4.5,3), key.title = NA, key.xlab="Z-score", key.ylab=NA,
          scale="none", symbreaks=T, symm=F, symkey = F,
          srtCol=45, adjCol=c(0.9,0), adjRow = c(0.5,0),
          cexRow = 0.8, cexCol = 0.8, sepwidth = c(0,0),
          main="Top 5% enriched")

p.neg <- z.p[, which(z.p[20,] < -20)]
heatmap.2(p.neg, col=my_palette, dendrogram='none',
          trace='none', Rowv=F, Colv=F, revC=T, #labCol=NA,
          xlab="Cellular Composition Terms", ylab="Quantiles of Tumor Cells Bayes Factors",
          margins = c(4.5,3), key.title = NA, key.xlab="Z-score", key.ylab=NA,
          scale="none", symbreaks=T, symm=F, symkey = F,
          srtCol=45, adjCol=c(0.9,0), adjRow=c(0.5,0),
          cexRow = 0.8, cexCol=0.8, sepwidth = c(0,0),
          main="Top 5% suppressed")

# Normal Cells
np.pos <- z.np[, which(z.np[20,] > 20)]
heatmap.2(np.pos, col=my_palette, dendrogram='none',
          trace='none', Rowv=F, Colv=F, revC=T, #labCol=NA,
          xlab="Cellular Composition Terms", ylab="Quantiles of Normal Cells Bayes Factors",
          margins = c(4.5,3), key.title = NA, key.xlab="Z-score", key.ylab=NA,
          scale="none", symbreaks=T, symm=F, symkey = F,
          srtCol=45, adjCol=c(0.9,0), adjRow=c(0.5,0),
          cexRow = 0.8, cexCol=0.8, sepwidth = c(0,0),
          main="Top 5% enriched")

np.neg <- z.np[, which(z.np[20,] < -20)]
heatmap.2(np.neg, col=my_palette, dendrogram='none',
          trace='none', Rowv=F, Colv=F, revC=T, #labCol=NA,
          xlab="Cellular Composition Terms", ylab="Quantiles of Normal Cells Bayes Factors",
          margins = c(4.5,3), key.title = NA, key.xlab="Z-score", key.ylab=NA,
          scale="none", symbreaks=T, symm=F, symkey = F,
          srtCol=45, adjCol=c(0.9,0), adjRow=c(0.5,0),
          cexRow = 0.8, cexCol=0.8, sepwidth = c(0,0),
          main="Top 5% suppressed")

# Differential z-scores
pvnp.pos <- z.pvnp[, which(z.pvnp[20,] > 4 & z.pvnp[1,] < z.pvnp[20,]/2 & z.pvnp[2,] < z.pvnp[20,]/2 & 
                                  z.pvnp[3,] < z.pvnp[20,]/2 & z.pvnp[4,] < z.pvnp[20,]/2 & 
                                  z.pvnp[5,] < z.pvnp[20,]/2 & z.pvnp[6,] < z.pvnp[20,]/2 & 
                                  z.pvnp[7,] < z.pvnp[20,]/2 & z.pvnp[8,] < z.pvnp[20,]/2 & 
                                  z.pvnp[9,] < z.pvnp[20,]/2 & z.pvnp[10,] < z.pvnp[20,]/2 &
                                  z.pvnp[11,] < z.pvnp[20,]-2 & z.pvnp[12,] < z.pvnp[20,]-2 &
                                  z.pvnp[13,] < z.pvnp[20,]-2 & z.pvnp[14,] < z.pvnp[20,]-2 &
                                  z.pvnp[15,] < z.pvnp[20,]-2 & z.pvnp[16,] < z.pvnp[20,]-2 & 
                                  z.pvnp[17,] < z.pvnp[20,]-2 & z.pvnp[18,] < z.pvnp[20,]-2 )]

png(filename = "17.PvsNP.z.cc.enriched.png",width=5.5, height=5.5, res=1200, unit="in")
heatmap.2(pvnp.pos, col=my_palette, dendrogram='none',
          trace='none', Rowv=F, Colv=F, revC=T, #labCol=NA,
          xlab="Cellular Composition Terms", ylab="Quantiles of Differential Fitness Score",
          margins = c(4.5,3), key.title = NA, key.xlab="Z-score", key.ylab=NA,
          scale="none", symbreaks=T, symm=F, symkey = F,
          srtCol=45, adjCol=c(0.9,0), adjRow=c(0.5,0),
          #lmat=rbind(c(0,3,4), c(2,1,0)), lwid=c(1.5,4,2),
          cexRow = 0.8, cexCol=0.8, sepwidth = c(0,0),
          main="Top 5% Enriched")
dev.off()

pvnp.neg <- z.pvnp[, which(z.pvnp[20,] < -4 & z.pvnp[1,] > z.pvnp[20,]/2 & z.pvnp[2,] > z.pvnp[20,]/2 & 
                                  z.pvnp[3,] > z.pvnp[20,]/2 & z.pvnp[4,] > z.pvnp[20,]/2 & 
                                  z.pvnp[5,] > z.pvnp[20,]/2 & z.pvnp[6,] > z.pvnp[20,]/2 & 
                                  z.pvnp[7,] > z.pvnp[20,]/2 & z.pvnp[8,] > z.pvnp[20,]/2 & 
                                  z.pvnp[9,] > z.pvnp[20,]/2 & z.pvnp[10,] > z.pvnp[20,]/2 &
                                  z.pvnp[11,] > z.pvnp[20,] & z.pvnp[12,] > z.pvnp[20,] &
                                  z.pvnp[13,] > z.pvnp[20,] & z.pvnp[14,] > z.pvnp[20,] &
                                  z.pvnp[15,] > z.pvnp[20,] & z.pvnp[16,] > z.pvnp[20,] & 
                                  z.pvnp[17,] > z.pvnp[20,] & z.pvnp[18,] > z.pvnp[20,] )]
png(filename = "17.PvsNP.z.cc.suppressed.png",width=5.5, height=5.5, res=1200, unit="in")
heatmap.2(pvnp.neg, col=my_palette, dendrogram='none',
          trace='none', Rowv=F, Colv=F, revC=T, #labCol= pvnp.neg.ids,
          xlab="Cellular Composition Terms", ylab="Quantiles of Differential Fitness Score",
          margins = c(4.5,3), key.title = NA, key.xlab="Z-score", key.ylab=NA,
          scale="none", symbreaks=T, symm=F, symkey = F,
          srtCol=45, adjCol=c(0.9,0), adjRow=c(0.5,0),
          #lmat=rbind(c(0,3,4), c(2,1,0)), lwid=c(1.5,4,2),
          cexRow = 0.8, cexCol=0.8, sepwidth = c(0,0),
          main="Top 5% Suppressed")
dev.off()

```
